requirements.txt
# Core Deep Learning & Scientific
torch
torchvision
torchaudio
timm
transformers
numpy

# NLP & Vision Specific
sentence-transformers
ultralytics
opencv-python
pillow

# Backend / Services
fastapi
uvicorn[standard]
python-multipart
kafka-python
redis

# Reporting & Utilities
pyyaml
python-pptx
tqdm

# Testing
pytest